chore: refactor style injection
