+++ lib/index.ts
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import ReactDOM, { unmountComponentAtNode } from 'react-dom'
 import getDisplayName from './getDisplayName'
+import { injectStylesBeforeElement } from './utils'
 
 function checkMountModeEnabled() {
   // @ts-ignore
@@ -11,50 +12,13 @@ function checkMountModeEnabled() {
   }
 }
 
-interface MountOptions {
-  alias?: string
-  ReactDom?: typeof ReactDOM
-  stylesheets?: string | string[]
-  style?: string
-  cssFile?: string
-}
-
 /**
  * Inject custom style text or CSS file or 3rd party style resources
  */
 const injectStyles = (options: MountOptions) => () => {
   const document = cy.state('document')
-
   const el = document.getElementById('cypress-jsdom')
-
-  // insert any custom global styles before the component
-  if (typeof options.stylesheets === 'string') {
-    options.stylesheets = [options.stylesheets]
-  }
-  if (Array.isArray(options.stylesheets)) {
-    // console.log('adding stylesheets')
-    options.stylesheets.forEach(href => {
-      const link = document.createElement('link')
-      link.type = 'text/css'
-      link.rel = 'stylesheet'
-      link.href = href
-      document.body.insertBefore(link, el)
-    })
-  }
-
-  if (options.style) {
-    const style = document.createElement('style')
-    style.appendChild(document.createTextNode(options.style))
-    document.body.insertBefore(style, el)
-  }
-
-  if (options.cssFile) {
-    return cy.readFile(options.cssFile).then(css => {
-      const style = document.createElement('style')
-      style.appendChild(document.createTextNode(css))
-      document.body.appendChild(style)
-    })
-  }
+  return injectStylesBeforeElement(options, document, el)
 }
 
 /**
